More recipe code cleaning
