+++ enderio-base/src/api/java/com/enderio/api/recipe/IEnderRecipe.java
@@ -1,11 +1,13 @@
 package com.enderio.api.recipe;
 
+import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
 import org.apache.commons.lang3.StringUtils;
 
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -19,52 +21,42 @@ public interface IEnderRecipe<R extends IEnderRecipe<R, C>, C extends Container>
         return true;
     }
 
-    default String getOwningMod() {
-        return getId().getNamespace();
-    }
-
-    /**
-     * Deprecated in favour of craft which supports multiple slots.
-     */
-    @Deprecated
     @Override
     default ItemStack assemble(C container) {
         throw new UnsupportedOperationException("assemble is disabled on custom recipes in favour of craft");
     }
 
-    //    List<ItemStack> craft(Container container);
-
     /**
      * Return a list of all possible inputs for each slot.
-     * Used by JEI primarily.
      */
     List<List<ItemStack>> getAllInputs();
 
     // TODO: Fluid ingredient support?
 
     /**
-     * Return a list of each slots output.
-     * Used by JEI primarily.
+     * Return a list of each slots' output.
      */
     List<ItemStack> getAllOutputs();
 
-    default List<ResourceLocation> getOtherDependencies() {
+    /**
+     * Get a list of other mod dependencies
+     */
+    default List<ResourceLocation> getMiscModDependencies() {
         return List.of();
     }
 
     default List<String> getModDependencies() {
+        // Generate a list of mod dependencies
         Set<String> mods = new HashSet<>();
-        // TODO!!!! Rework mod dependency system for inputs.
-        //        for (EIOIngredient ingredient : getInputs()) {
-        //            for (ItemStack item : ingredient.getItems()) {
-        //                String mod = item.getItem().getRegistryName().getNamespace();
-        //                mods.add(mod);
-        //            }
-        //        }
 
+        // Get mods for inputs, outputs and other mods
+        getIngredients().forEach(ing -> Arrays.stream(ing.getItems()).forEach(itm -> mods.add(itm.getItem().getRegistryName().getNamespace())));
         getAllOutputs().forEach(dep -> mods.add(dep.getItem().getRegistryName().getNamespace()));
-        getOtherDependencies().forEach(dep -> mods.add(dep.getNamespace()));
+        getMiscModDependencies().forEach(dep -> mods.add(dep.getNamespace()));
 
-        return mods.stream().filter(mod -> !StringUtils.equalsAny(mod, "minecraft", "forge", "enderio", getOwningMod())).toList();
+        // Get which mod owns the recipe type. Makes sure the mod that deserializes the recipe is ignored in the list.
+        String owningMod = Registry.RECIPE_TYPE.getKey(getType()).getNamespace();
+
+        return mods.stream().filter(mod -> !StringUtils.equalsAny(mod, "minecraft", "forge", "enderio", owningMod)).toList();
     }
 }
