+++ enderio-machines/src/main/java/com/enderio/machines/common/init/MachineRecipes.java
@@ -5,6 +5,8 @@ import com.enderio.api.recipe.AlloySmeltingRecipe;
 import com.enderio.machines.common.recipe.AlloySmeltingRecipeImpl;
 import com.enderio.api.recipe.EnchanterRecipe;
 import com.enderio.machines.common.recipe.EnchanterRecipeImpl;
+import com.enderio.machines.common.recipe.serializer.AlloySmeltingRecipeSerializer;
+import com.enderio.machines.common.recipe.serializer.EnchanterRecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -19,8 +21,8 @@ public class MachineRecipes {
 
         public static final DeferredRegister<RecipeSerializer<?>> RECIPE_SERIALIZER_REGISTRY = DeferredRegister.create(ForgeRegistries.RECIPE_SERIALIZERS, EIOMachines.MODID);
 
-        public static final RegistryObject<EnchanterRecipeImpl.Serializer> ENCHANTING = RECIPE_SERIALIZER_REGISTRY.register("enchanting", EnchanterRecipeImpl.Serializer::new);
-        public static final RegistryObject<AlloySmeltingRecipeImpl.Serializer> ALLOY_SMELTING = RECIPE_SERIALIZER_REGISTRY.register("alloy_smelting", AlloySmeltingRecipeImpl.Serializer::new);
+        public static final RegistryObject<EnchanterRecipeSerializer> ENCHANTING = RECIPE_SERIALIZER_REGISTRY.register("enchanting", EnchanterRecipeSerializer::new);
+        public static final RegistryObject<AlloySmeltingRecipeSerializer> ALLOY_SMELTING = RECIPE_SERIALIZER_REGISTRY.register("alloy_smelting", AlloySmeltingRecipeSerializer::new);
 
         public static void classload() {
             RECIPE_SERIALIZER_REGISTRY.register(FMLJavaModLoadingContext.get().getModEventBus());
